pyproject.toml
[tool.poetry]
name = "ai-agent"
version = "0.1.0"
description = "AI coding agent based on Learn-by-interact principles"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.0.0"
click = "^8.1.7"
rich = "^13.0.0"
gitpython = "^3.1.0"
numpy = "^1.26.0"
pytest = "^7.0.0"
typing-extensions = "^4.0.0"
rank-bm25 = "^0.2.2"
sentence-transformers = "^2.2.2"

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
isort = "^5.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ai-agent = "ai_agent.cli:main"